github: include the instructions as comments in the templates (#47)
Use Markdown comments for the instructions in the PR and issues
templates. This way you don't need to remove the instructions as you
fill out the form, but they aren't rendered in the resulting issue/PR.

## How to use

Create issues and PRs without needing to delete the instructions as you
go.

## Testing done

Examined the output in the VSCode markdown preview window to see the
comments not get rendered. This is also the [documented
syntax](https://github.github.com/gfm/#html-comment).