Let the user provide the reqwest Client to libazureinit calls
This allows the user to configure and re-use a reqwest client for HTTP
requests issued by the library. This way they can set default headers,
timeout behavior, benefit from connection pooling, etc.

As part of this change, a client is configured in the azure-init binary
that sets a user-agent and a request timeout. Without the request
timeout, the client can wait indefinitely.

Fixes #54