Use safe nix interface to retrieve user and group id
Using libc APIs directly is difficult and should be avoided if possible.
In this case, the manual for getpwnam(3) and getgrnam(3) indicate they
return NULL if the entry is not found or an error occurs. Since the
return value isn't being checked for NULL here, a null pointer
dereference can occur here.

Fortunately, the nix crate provides a wrapper for retrieving user
details from a username.

I've added some calls to unwrap() which I'll remove once more error
handling is in place.