Implement a builder-style provisioning API
As discussed in PR #86, this is a builder-style interface to allow
library users to select the backend for tools when provisioning.

Library users must specify one or more provisioners for each setting.
This allows users to decide which tool or tools to use when
provisioning. Some feature flags have been added to `azure-init` which
enable provisioning with a tool, letting you build binaries for a
particular platform relatively easily.