.github: migrate to dtolnay/rust-toolchain
Migrate to GitHub Actions dtolnay/rust-toolchain, to support multiple
toolchains as well as an old toolchain like stable 12 months ago.

Note: it is not possible to go back to 18 months ago, Rust 1.66.1 as of
now, because then nix v0.29.0 of the current code base cannot be built,
nix 0.29.0 requires rustc 1.69 or newer.