Get username either by IMDS or from a local OVF file
Username can be obtained either via fetching instance metadata from IMDS
or mounting a local device for OVF environment file. It should not fail
immediately in a single failure, instead it should fall back to the other
mechanism. So it is not a good idea to use `?` for query() or
get_environment().

Explicitly get instance metadata to handle error of missing instance metadata
before dealing with setting ssh keys or hostname.