Get username either by IMDS or from a local OVF file (#100)
* Get username either by IMDS or from a local OVF file

Username can be obtained either via fetching instance metadata from IMDS
or mounting a local device for OVF environment file. It should not fail
immediately in a single failure, instead it should fall back to the other
mechanism. So it is not a good idea to use `?` for query() or
get_environment().

Explicitly get instance metadata to handle error of missing instance metadata
before dealing with setting ssh keys or hostname.

* tests: pass in security type when creating VM

When running functional tests, pass in --security-type, usually
`TrustedLaunch`, which is nowadays the default for most VMs.
It is still possible to override the security type by passing in
an environment variable $VM_SECURITY_TYPE, like `Standard`.

* tests: address shellcheck warnings in functional_tests.sh