Add an initial mkosi config
This is still a work-in-progress.

It still needs:
 - A stubbed out Wireserver and IMDS instance

To try it out as-is, you can build and then boot the image with:

$ mkosi -d fedora
$ sudo mkosi boot

The version of azure-init in target/debug/ will be run as part of the
image startup. You can inspect the result with "systemctl status
azure-init". Run "systemctl poweroff" to leave the environment.

This needs to be done in an Azure VM since it relies on talking to a
real Wireserver and IMDS.

Refer to mkosi docs[0] for installation options

[0] https://github.com/systemd/mkosi?tab=readme-ov-file#installation