fix(ssh): use sshd -G over -T and add error handling
sshd -T requires sshd to be running which it may not be.
This results in failure to provision correctly.

Update module to use sshd -G instead.

Since -G is only supported in newer versions of OpenSSH,
ensure we default to .ssh/authorized_keys on failure.

As a bonus, -G appears more performant that -T in my
experiments.

- Add authorized_keys_path_string parameter to provision_ssh()
  for dependency injection.  This allows removal of mock_provision_ssh()
  and using the real function without the sshd invocation.

- Add get_authorized_keys_path_from_sshd() with parameter
  for runner.  This allows us to test the various failure modes
  when invoking sshd -G separately from provision_ssh() tests.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>