Refactor the HTTP retry code to use a single implementation (#135)
* error: include the underlying error in the error text

While these may not be the optimal error messages in all cases, this
should make them more useful by including the underlying error message
so that you get, for example, "HTTP client error occurred: HTTP status
client error (405 Method Not Allowed) for url (http://127.0.0.1:35195/)"
instead of "HTTP client error ocurred".

* http: provide a generic HTTP client interface that retries

Add a generic `request()` function which implements the retry rules we
have and provide `get()` and `post()` wrappers for the function. Move
the IMDS and Wireserver interfaces to this new interface so we don't
have multiple implementations of the retry rules.

* Retry when the response body can't be deserialized

In the event we receive something that can't be deserialized, retry the
request.

This seems like a rather unlikely scenario, since I would expect a
partial response to fail with a lower-level reqwest error, but it can
happen in theory, so let's handle it gracefully.

* imds: bump API version and set extended=true to the query parameters

This causes the request to be treated as a request for data for
provisioning purposes. Without it, the server may respond with an HTTP
200 and the response may not include all the data we need to provision.

Using this query parameter should turn those requests into failures
which we'll retry.

Fixes #134