Standardize logging when running subprocesses
This adds a helper function which accepts a Command to run and has
uniform logging when running a command.

The command is logged at the debug level prior to execution, and
afterwards the status is also logged at debug level. If the command
failed, it's logged at error level and includes the command's stdout and
stderr.

This also ensures all commands run used piped stdio rather than
inheriting the parent. Allowing the child to write to the parent
stdout/stderr leads to unstructured command output mixed into the stderr
log output. The easiest way to do this is to use `output()` rather than
`status()` when running the command.